Genericize EnumVariant
